requirements.txt
aio-pika==9.2.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiormq==6.7.7
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
markdownify==0.11.6
multidict==6.0.4
pamqp==3.2.1
pika==1.3.2
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-json-logger==2.0.7
pytz==2023.3
redis==4.6.0
requests==2.31.0
sentry-sdk==1.29.2
six==1.16.0
soupsieve==2.4.1
twilio==8.5.0
urllib3==2.0.4
yarl==1.9.2